Revise app file structure
